requirements.txt
Mastodon.py
requests
psutil
